refactor(providers): land the three Phase 8 conformance preconditions
Part of #860 (Phase 8), closing the preconditions recorded in #974:

- optional_api_key discovery cells now derive auth: 'optional' and the
  generated matrix executes both credential branches (configured key
  reaches the wire; absent key sends no credential at all).
- Override cells now bind to executable runs in
  provider-contract-overrides.ts that the matrix suite executes
  directly, replacing the textual test-title source check; the bound
  hand-written tests moved out of provider-conformance.test.ts /
  model-fetcher.test.ts so each contract runs exactly once.
- The /^gpt-5/i Responses routing regex is lifted into the
  openAiAdapterApiProtocol declaration in core (ModelInfo.apiProtocol
  seam); runtime model-factory and the matrix sample-id derivation both
  consume the declaration, with a wire test locking gpt-5*->Responses
  and other ids->Chat Completions.
