+++ packages/runtime/src/__tests__/conformance-harness.ts
@@ -1,0 +1,62 @@
+/**
+ * Shared scripted-HTTP harness for the provider conformance suites.
+ *
+ * Not a test file: node --test only picks up `*.test.js`, so this module can be
+ * imported by the generated matrix suite, the executable override bindings, and
+ * the hand-written conformance suite without registering anything itself.
+ *
+ * Servers are tracked per process; each suite that starts servers must call
+ * {@link closeAllJsonServers} from an `after` hook.
+ */
+
+import assert from 'node:assert/strict';
+import { createServer, type IncomingMessage, type ServerResponse } from 'node:http';
+
+const servers: Array<{ close(): Promise<void> }> = [];
+
+export async function startJsonServer(
+  handler: (request: IncomingMessage, response: ServerResponse) => void | Promise<void>,
+): Promise<{ url: string; close(): Promise<void> }> {
+  const server = createServer((request, response) => {
+    void Promise.resolve(handler(request, response)).catch((error) => {
+      response.destroy(error as Error);
+    });
+  });
+  await new Promise<void>((resolve) => server.listen(0, '127.0.0.1', resolve));
+  const address = server.address();
+  assert.ok(address && typeof address === 'object');
+  const control = {
+    url: `http://127.0.0.1:${address.port}`,
+    close: () => new Promise<void>((resolve, reject) => {
+      server.close((error) => (error ? reject(error) : resolve()));
+    }),
+  };
+  servers.push(control);
+  return control;
+}
+
+export async function closeAllJsonServers(): Promise<void> {
+  const open = servers.splice(0, servers.length);
+  await Promise.all(open.map((server) => server.close()));
+}
+
+export function readBody(request: IncomingMessage): Promise<string> {
+  return new Promise((resolve, reject) => {
+    let body = '';
+    request.setEncoding('utf8');
+    request.on('data', (chunk) => { body += chunk; });
+    request.on('end', () => resolve(body));
+    request.on('error', reject);
+  });
+}
+
+export function respondJson(response: ServerResponse, status: number, body: unknown): void {
+  response.writeHead(status, { 'content-type': 'application/json' });
+  response.end(JSON.stringify(body));
+}
+
+export function respondOpenAIStream(response: ServerResponse, chunks: readonly unknown[]): void {
+  response.writeHead(200, { 'content-type': 'text/event-stream' });
+  for (const chunk of chunks) response.write(`data: ${JSON.stringify(chunk)}\n\n`);
+  response.end('data: [DONE]\n\n');
+}
